==> 1.b.tokenization_topics.ipynb <==
import nltk
import pandas as pd
import numpy as np
import pickle

from nltk.tokenize import word_tokenize
from nltk.corpus import stopwords
from sklearn.feature_extraction.text import CountVectorizer

### Load Data
#### Load "derecha diario" dataset from csv.

df = pd.read_csv('./data/derecha_econ_news.csv', index_col=0)

df.shape

df.head(5)

# remove unicode
df.body.replace('\n', ' ', regex=True ,inplace=True)
df.body.replace('\xa0', ' ', regex=True ,inplace=True)

### Topics
Selection of topics from news titles of the portal's economic section.

inflation = df.loc[df.title.str.contains('inflacion|inflación|emisión|salarios|salario|indec', na=False)]
inflation.shape

exchange = df.loc[df.title.str.contains('dolar|dólar|dolares|dólares|dolarización|cambio|cepo', na=False)]
exchange.shape

### Corpus (Body)
Clipping of the selected news body from 2020 and erasing nulls

#### Topic: Inflation

year = inflation.date.str.contains('2020', na=False)
year.sum()

inflation_2020 = inflation[year]
inflation_2020.dropna(inplace=True)
inflation_2020.shape

inflation_2020 = inflation_2020.reset_index()

inflation_2020.insert(loc=1, column='topics', value='inflation')

inflation_2020.head(5)

#### Topic: Currency Exchange

year = exchange.date.str.contains('2020', na=False)
year.sum()

exchange_2020 = exchange[df.date.str.contains('2020', na=False)]
exchange_2020.dropna(inplace=True)
exchange_2020.shape

exchange_2020 = exchange_2020.reset_index()

exchange_2020.insert(loc=1, column='topics', value='exchange')

exchange_2020.head(5)

data_clean = inflation_2020.append(exchange_2020)
data_clean.shape

I drop those news labeled under both topics

data_clean = data_clean.drop_duplicates(subset='index',keep=False)
data_clean.shape

Check the number of words per article's body and discard those incompleted

filter = data_clean.body.apply(lambda x: len(word_tokenize(x))) < 150
filter.sum()

data_clean.drop(data_clean[filter].index, axis=0, inplace=True)

#### Save selection

data_clean.reset_index(drop=True, inplace=True)

data_clean.drop('index', axis=1, inplace=True)

print('num inflation news:', data_clean.loc[data_clean.topics == 'inflation',].shape[0])
print('num exchange news:', data_clean.loc[data_clean.topics == 'exchange',].shape[0])

data_clean.sample(5)

pickle.dump(data_clean, open( "data/df_clean_der.pkl", "wb" ))

corpus = data_clean
corpus.shape

### Tokenization
#### Manual

import re
import unidecode

"Stop words", special characters, accents and numbers are removed

stopwords_sp = stopwords.words('spanish')

def tokenizer(texto):

    alphanumeric = re.sub(r'([^\s\w]|_)+', '', texto).lower()
    no_accents = unidecode.unidecode(alphanumeric)
    
    tockens = word_tokenize(no_accents)
    
    tockens_clean = [tocken for tocken in tockens if tocken not in stopwords_sp and tocken.isalpha()]
    
    terminos = tockens_clean

    return terminos

tokenized = [tokenizer(documento) for documento in corpus.body]

vocabulario = sorted(list(set([word for group in tokenized for word in group])))

len(vocabulario)

#### CountVectorizer

from sklearn.feature_extraction.text import CountVectorizer

Removal of special characters function to load in vectorizer

def removal(text):
    text = re.sub(r'(\d|\$|\%|\+)', '', text.lower())
    return re.sub(r'\d+', '', text)

vectorizer = CountVectorizer(stop_words = stopwords_sp, lowercase = True, strip_accents='unicode', preprocessor=removal)

vectorizer.fit(corpus.body)

len(vectorizer.vocabulary_), len(vectorizer.get_feature_names())

I prefare the manual tokenization results since the process is clearer and there are more terms.

### Pickle

Manual tokens

def token_matrix_maker(data, vocabulario):
    matriz = np.zeros(shape = (len(data), len(vocabulario)), dtype='int')
    for i, documento in enumerate(data):
        for termino in documento:
            matriz[i, vocabulario.index(termino)] += 1
    return matriz

matrix = token_matrix_maker(tokenized, vocabulario)

# make dataframe
token_df = pd.DataFrame(matrix, columns=vocabulario, index = ['doc' + str(i + 1) for i in range(len(tokenized))] )

# add classification column
topics = ['inflation' for i in range((corpus.topics == 'inflation').sum())] + ['exchange' for i in range((corpus.topics == 'exchange').sum())]
token_df.insert(loc=0, column='topics', value=topics)

token_df

token_df.to_pickle('tokens/topic_tokens_der.pkl')

#### CV tokens

pickle.dump(vectorizer, open( "tokens/cv_der.pkl", "wb" ))
