==> 4.a.naive_topics.ipynb <==
## Naive Bayes
Inflation VS currency exchange in "Izquierda Diario"

import pandas as pd
import pickle

data = pickle.load(open('data/df_clean_izq.pkl','rb'))

data.sample(5)

df_inflation = data.loc[data.topics == 'inflation']
df_inflation.shape

df_exchange = data.loc[data.topics == 'exchange']
df_exchange.shape

corpus = list(df_inflation.body.values) + list(df_exchange.body.values)
len(corpus)

target = list(df_inflation.topics.values) + list(df_exchange.topics.values)

data = pd.DataFrame([corpus, target]).transpose()
data

### Train - Test Split

from sklearn.model_selection import train_test_split

train,test,y_train,y_test = train_test_split(data.iloc[:,0], data.iloc[:,1], train_size=0.75); 

train.shape, test.shape

### Tokenization + Vectorization
#### CountVectorizer

from sklearn.feature_extraction.text import CountVectorizer
from nltk.corpus import stopwords
import re

stopwords_sp = stopwords.words('spanish')

vectorizer = CountVectorizer(stop_words = stopwords_sp, lowercase = True, strip_accents='unicode')
vectorizer.fit(list(re.sub(r'(\d|\$|\%|\+)', '', doc) for doc in corpus))

X_train = vectorizer.transform(train);

X_train

X_train_matriz = X_train.todense()

X_test = vectorizer.transform(test);

X_test_matriz = X_train.todense()

#### Tf IDf Vectorizer

from sklearn.feature_extraction.text import TfidfTransformer

X_train_matriz_tfidf = TfidfTransformer().fit_transform(X_train_matriz);

X_test_matriz_tfidf = TfidfTransformer().fit_transform(X_test_matriz);

### Multinomial Naive Bayes + CountVectorizer

from sklearn.naive_bayes import MultinomialNB
from sklearn.metrics import accuracy_score
from sklearn.metrics import confusion_matrix

multi_nb = MultinomialNB()

multi_nb.fit(X_train, y_train.values)

y_pred = multi_nb.predict(X_test)

from sklearn.metrics import accuracy_score

accuracy_score(y_pred, y_test.values)

confusion_matrix(y_pred, y_test.values)

### Multinomial Naive Bayes + TfIDf

multi_nb_2 = MultinomialNB()

multi_nb_2.fit(X_train_matriz_tfidf, y_train.values)

y_pred = multi_nb_2.predict(X_test)

from sklearn.metrics import accuracy_score

accuracy_score(y_pred, y_test.values)

confusion_matrix(y_pred, y_test.values)

### Tokenization + Vectorization (Manual)

#### CountVectorizer with pre-built tokens

cv_tokens = pickle.load(open('tokens/topic_tokens_izq_new.pkl', 'rb'))

vectorizer.fit(cv_tokens.drop('topics', axis=1))

X_train_2 = vectorizer.transform(train)

X_train_matriz_2 = X_train_2.todense()

X_test_2 = vectorizer.transform(test)

X_test_matriz_2 = X_test_2.todense()

#### TF-IDF Vectorizer with pre-built tokens

X_train_matriz_tfidf_2 = TfidfTransformer().fit_transform(X_train_matriz_2);

X_test_matriz_tfidf_2 = TfidfTransformer().fit_transform(X_test_matriz_2);

### Multinomial Naive Bayes + CountVectorizer (updated tokens)

from sklearn.naive_bayes import MultinomialNB
from sklearn.metrics import accuracy_score
from sklearn.metrics import confusion_matrix

multi_nb = MultinomialNB()

multi_nb.fit(X_train_2, y_train.values)

y_pred_2 = multi_nb.predict(X_test_2)

from sklearn.metrics import accuracy_score

accuracy_score(y_pred_2, y_test.values)

confusion_matrix(y_pred_2, y_test.values)

### Multinomial Naive Bayes + Tf IDf Vectorizer (updated tokens)

multi_nb_2 = MultinomialNB()

multi_nb_2.fit(X_train_matriz_tfidf_2, y_train.values)

y_pred_2 = multi_nb_2.predict(X_test_matriz_tfidf_2)

from sklearn.metrics import accuracy_score

accuracy_score(y_pred_2, y_test.values)

confusion_matrix(y_pred_2, y_test.values)
