==> 7.a.vs_model_sel_topic_1.ipynb <==
## Vectorizers + Models
#### Inflation topic

### Load

Tokens

import warnings
warnings.filterwarnings('ignore')

import pandas as pd
import numpy as np

tokens_izq = pd.read_pickle('tokens/topic_tokens_izq_new.pkl')
tokens_der = pd.read_pickle('tokens/topic_tokens_der_new.pkl')

tokens_izq_inf = tokens_izq[tokens_izq.topics == 'inflation'].drop('topics', axis=1)
tokens_izq_inf.insert(loc=0, column='publication', value='left')

tokens_der_inf = tokens_der[tokens_der.topics == 'inflation'].drop('topics', axis=1)
tokens_der_inf.insert(loc=0, column='publication', value='right')

tokens = tokens_izq_inf.append(tokens_der_inf)
tokens.reset_index(inplace=True)
tokens.drop('index', axis=1, inplace=True)
tokens.fillna(0, inplace=True)

tokens.sample(5)

#### Train - Test Split

from sklearn.model_selection import train_test_split

# data balance
tokens.publication.value_counts(normalize=True)

X = tokens.drop('publication', axis=1)

y = tokens['publication']

train, test, y_train, y_test = train_test_split(X, y, train_size=0.75, stratify=tokens.publication)

Shapes and balance

# shape train
display(train.shape[0],
        train.shape[0] / X.shape[0])

# shape test
display(test.shape[0],
        test.shape[0] / X.shape[0])

# data balance - test
display(y_train.value_counts(normalize=True), 
        y_test.value_counts(normalize=True))

Vectorización con TF-IDF

from sklearn.feature_extraction.text import TfidfTransformer

tfidf_vector = TfidfTransformer().fit(train)
train_tfidf = tfidf_vector.transform(train)

test_tfidf = tfidf_vector.transform(test)

#### Models calling

from sklearn.naive_bayes import MultinomialNB

from sklearn.linear_model import LogisticRegression, SGDClassifier

from sklearn.neighbors import KNeighborsClassifier

from sklearn.tree import DecisionTreeClassifier

from sklearn.ensemble import AdaBoostClassifier, GradientBoostingClassifier

from xgboost import XGBClassifier

#### K-Folds + Cross Validation

from sklearn.model_selection import StratifiedKFold, cross_val_score

from sklearn.metrics import accuracy_score

folds = StratifiedKFold(n_splits=5, random_state=19, shuffle=True)

estimators = [
    ('MNB', MultinomialNB()),
    ('LR', LogisticRegression()),
    ('SGD', SGDClassifier()),
    ('KNN', KNeighborsClassifier()),
    ('CART', DecisionTreeClassifier()),
    ('ADA', AdaBoostClassifier()),
    ('GDB', GradientBoostingClassifier()),
    ('XGB', XGBClassifier())
]
results = []

results = []
scoring = 'accuracy'
print("model\tCV mean\t CV std")
for name, model in estimators:
 	cv_results = cross_val_score(model, X, y, cv=folds, scoring='accuracy')
 	results.append(cv_results)
 	msg = "%s:\t%f (%f)" % (name, cv_results.mean(), cv_results.std())
 	print(msg)

### GridSearchCV (Vectorization + Models)

from sklearn.model_selection import GridSearchCV

### 1. Naive Bayes

CountVectorizer

parameters = {
        'alpha': (1e-2, 1e-3, 1e-1),
    }

grid_mnb = GridSearchCV(MultinomialNB(), parameters, cv=folds, scoring='accuracy')

grid_mnb.fit(train, y_train)

mnb_train_score_cv = grid_mnb.score(train, y_train)
mnb_test_score_cv = accuracy_score(grid_mnb.predict(test), y_test)

mnb_cv_fold_test_mean = grid_mnb.cv_results_['mean_test_score'].mean()
mnb_cv_fold_test_std = grid_mnb.cv_results_['std_test_score'].mean()

print(f'''Best parameters: {grid_mnb.best_params_}
Mean cv test score: {grid_mnb.cv_results_['mean_test_score'].mean()}
Std cv test score: {grid_mnb.cv_results_['std_test_score'].mean()}
Train score: {mnb_train_score_cv}
Test test score: {mnb_test_score_cv}''')

Tfidfvectorizer

grid_mnb.fit(train_tfidf, y_train)

mnb_train_score_td = grid_mnb.score(train_tfidf, y_train)
mnb_test_score_td = accuracy_score(grid_mnb.predict(test_tfidf), y_test)

mnb_td_fold_test_mean = grid_mnb.cv_results_['mean_test_score'].mean()
mnb_td_fold_test_std = grid_mnb.cv_results_['std_test_score'].mean()

print(f'''Best parameters: {grid_mnb.best_params_}
Mean cv test score: {grid_mnb.cv_results_['mean_test_score'].mean()}
Std cv test score: {grid_mnb.cv_results_['std_test_score'].mean()}
Train score: {mnb_train_score_td}
Test test score: {mnb_test_score_td}''')

### 2. LogisticRegression

CountVectorizer

parameters = {
        'C': [1, 10, 100, 1000],
        'penalty': ['l1', 'l2',],
        'solver': ['saga']
    }

grid_log = GridSearchCV(LogisticRegression(), parameters, cv=folds, scoring='accuracy')

grid_log.fit(train, y_train)

log_cv_train_score = grid_log.score(train, y_train)
log_cv_test_score = accuracy_score(grid_log.predict(test), y_test)

log_cv_fold_test_mean = grid_log.cv_results_['mean_test_score'].mean()
log_cv_fold_test_std = grid_log.cv_results_['std_test_score'].mean()

print(f'''Best parameters: {grid_log.best_params_}
Mean cv test score: {grid_log.cv_results_['mean_test_score'].mean()}
Std cv test score: {grid_log.cv_results_['std_test_score'].mean()}
Train score: {log_cv_train_score}
Test test score: {log_cv_test_score}''')

Tfidfvectorizer

grid_log.fit(train_tfidf, y_train)

log_td_train_score = grid_log.score(train_tfidf, y_train)
log_td_test_score = accuracy_score(grid_log.predict(test_tfidf), y_test)

log_td_fold_test_mean = grid_log.cv_results_['mean_test_score'].mean()
log_td_fold_test_std = grid_log.cv_results_['std_test_score'].mean()

print(f'''Best parameters: {grid_log.best_params_}
Mean cv test score: {grid_log.cv_results_['mean_test_score'].mean()}
Std cv test score: {grid_log.cv_results_['std_test_score'].mean()}
Train score: {log_td_train_score}
Test test score: {log_td_test_score}''')

### 3. SGDClassifier

CountVectorizer

parameters ={
    'penalty': ('l2', 'elasticnet', 'l1'),
    'max_iter': [50, 80],
    'tol': [1e-4],
    'loss': ['hinge', 'log', 'modified_huber'],
}

grid_sgd = GridSearchCV(SGDClassifier(), parameters, cv=folds, scoring='accuracy')

grid_sgd.fit(train, y_train)

sgd_cv_train_score = grid_sgd.score(train, y_train)
sgd_cv_test_score = accuracy_score(grid_sgd.predict(test), y_test)

sgd_cv_fold_test_mean = grid_sgd.cv_results_['mean_test_score'].mean()
sgd_cv_fold_test_std = grid_sgd.cv_results_['std_test_score'].mean()

print(f'''Best parameters: {grid_sgd.best_params_}
Mean cv test score: {grid_sgd.cv_results_['mean_test_score'].mean()}
Std cv test score: {grid_sgd.cv_results_['std_test_score'].mean()}
Train score: {sgd_cv_train_score}
Test test score: {sgd_cv_test_score}''')

Tfidfvectorizer

grid_sgd.fit(train_tfidf, y_train)

sgd_td_train_score = grid_sgd.score(train_tfidf, y_train)
sgd_td_test_score = accuracy_score(grid_sgd.predict(test_tfidf), y_test)

sgd_td_fold_test_mean = grid_sgd.cv_results_['mean_test_score'].mean()
sgd_td_fold_test_std = grid_sgd.cv_results_['std_test_score'].mean()

print(f'''Best parameters: {grid_sgd.best_params_}
Mean cv test score: {grid_sgd.cv_results_['mean_test_score'].mean()}
Std cv test score: {grid_sgd.cv_results_['std_test_score'].mean()}
Train score: {log_td_train_score}
Test test score: {log_td_test_score}''')

### 4. KNeighborsClassifier

CountVectorizer

parameters ={
        'n_neighbors': range(1,4),
        'weights' : ['uniform', 'distance'],
        'p' : [1, 2, 3]
}

grid_knn = GridSearchCV(KNeighborsClassifier(), parameters, cv=folds, scoring='accuracy')

grid_knn.fit(train, y_train)

knn_cv_train_score = grid_knn.score(train, y_train)
knn_cv_test_score = accuracy_score(grid_knn.predict(test), y_test)

knn_cv_fold_test_mean = grid_knn.cv_results_['mean_test_score'].mean()
knn_cv_fold_test_std = grid_knn.cv_results_['std_test_score'].mean()

print(f'''Best parameters: {grid_knn.best_params_}
Mean cv test score: {grid_knn.cv_results_['mean_test_score'].mean()}
Std cv test score: {grid_knn.cv_results_['std_test_score'].mean()}
Train score: {knn_cv_train_score}
Test test score: {knn_cv_test_score}''')

Tfidfvectorizer

grid_knn.fit(train_tfidf, y_train)

knn_td_train_score = grid_knn.score(train_tfidf, y_train)
knn_td_test_score = accuracy_score(grid_knn.predict(test_tfidf), y_test)

knn_td_fold_test_mean = grid_knn.cv_results_['mean_test_score'].mean()
knn_td_fold_test_std = grid_knn.cv_results_['std_test_score'].mean()

print(f'''Best parameters: {grid_knn.best_params_}
Best cv score: {grid_knn.best_score_}
Train score: {knn_td_train_score}
Test test score: {knn_td_test_score}''')

### 5. CART

CountVectorizer

parameters ={
        "criterion" : ["gini", "entropy"],
        "min_samples_leaf": [5,10,15,20,2], 
        "max_depth" : [1,2,3,4,5,6,8,9,10,11,12,13,14,15,16,17],
        "min_samples_split": [2, 3, 4,None]
}

grid_dtc = GridSearchCV(DecisionTreeClassifier(), parameters, cv=folds, scoring='accuracy')

grid_dtc.fit(train, y_train)

dtc_cv_train_score = grid_dtc.score(train, y_train)
dtc_cv_test_score = accuracy_score(grid_dtc .predict(test), y_test)

dtc_cv_fold_test_mean = grid_dtc.cv_results_['mean_test_score'].mean()
dtc_cv_fold_test_std = grid_dtc.cv_results_['std_test_score'].mean()

print(f'''Best parameters: {grid_dtc.best_params_}
Mean cv test score: {grid_dtc.cv_results_['mean_test_score'].mean()}
Std cv test score: {grid_dtc.cv_results_['std_test_score'].mean()}
Train score: {dtc_cv_train_score}
Test test score: {dtc_cv_test_score}''')

Tfidfvectorizer

grid_dtc.fit(train_tfidf, y_train)

dtc_td_train_score = grid_dtc.score(train_tfidf, y_train)
dtc_td_test_score = accuracy_score(grid_dtc.predict(test_tfidf), y_test)

dtc_td_fold_test_mean = grid_dtc.cv_results_['mean_test_score'].mean()
dtc_td_fold_test_std = grid_dtc.cv_results_['std_test_score'].mean()

print(f'''Best parameters: {grid_dtc.best_params_}
Mean cv test score: {grid_dtc.cv_results_['mean_test_score'].mean()}
Std cv test score: {grid_dtc.cv_results_['std_test_score'].mean()}
Train score: {knn_td_train_score}
Test test score: {knn_td_test_score}''')

### 6. AdaBoost

CountVectorizer

parameters = {
    'learning_rate': [.5]
}

grid_ada = GridSearchCV(AdaBoostClassifier(), parameters, cv=folds, scoring='accuracy')

grid_ada.fit(train, y_train)

ada_cv_train_score = grid_ada.score(train, y_train)
ada_cv_test_score = accuracy_score(grid_ada .predict(test), y_test)

ada_cv_fold_test_mean = grid_ada.cv_results_['mean_test_score'].mean()
ada_cv_fold_test_std = grid_ada.cv_results_['std_test_score'].mean()

print(f'''Best parameters: {grid_ada.best_params_}
Mean cv test score: {grid_ada.cv_results_['mean_test_score'].mean()}
Std cv test score: {grid_ada.cv_results_['std_test_score'].mean()}
Train score: {ada_cv_train_score}
Test test score: {ada_cv_test_score}''')

Tfidfvectorizer

grid_ada.fit(train_tfidf, y_train)

ada_td_train_score = grid_ada.score(train_tfidf, y_train)
ada_td_test_score = accuracy_score(grid_ada.predict(test_tfidf), y_test)

ada_td_fold_test_mean = grid_ada.cv_results_['mean_test_score'].mean()
ada_td_fold_test_std = grid_ada.cv_results_['std_test_score'].mean()

print(f'''Best parameters: {grid_ada.best_params_}
Mean cv test score: {grid_ada.cv_results_['mean_test_score'].mean()}
Std cv test score: {grid_ada.cv_results_['std_test_score'].mean()}
Train score: {ada_td_train_score}
Test test score: {ada_td_test_score}''')

### 7. GradientBoosting

CountVectorizer

parameters = {
    'loss':['deviance', 'exponential']
}

grid_grb = GridSearchCV(GradientBoostingClassifier(), parameters, cv=folds, scoring='accuracy')

grid_grb.fit(train, y_train)

grb_cv_train_score = grid_grb.score(train, y_train)
grb_cv_test_score = accuracy_score(grid_grb.predict(test), y_test)

grb_cv_fold_test_mean = grid_grb.cv_results_['mean_test_score'].mean()
grb_cv_fold_test_std = grid_grb.cv_results_['std_test_score'].mean()

print(f'''Best parameters: {grid_grb.best_params_}
Mean cv test score: {grid_grb.cv_results_['mean_test_score'].mean()}
Std cv test score: {grid_grb.cv_results_['std_test_score'].mean()}
Train score: {grb_cv_train_score}
Test test score: {grb_cv_test_score}''')

Tfidfvectorizer

grid_grb.fit(train_tfidf, y_train)

grb_td_train_score = grid_grb.score(train_tfidf, y_train)
grb_td_test_score = accuracy_score(grid_grb.predict(test_tfidf), y_test)

grb_td_fold_test_mean = grid_grb.cv_results_['mean_test_score'].mean()
grb_td_fold_test_std = grid_grb.cv_results_['std_test_score'].mean()

print(f'''Best parameters: {grid_grb.best_params_}
Mean cv test score: {grid_grb.cv_results_['mean_test_score'].mean()}
Std cv test score: {grid_grb.cv_results_['std_test_score'].mean()}
Train score: {grb_td_train_score}
Test test score: {grb_td_test_score}''')

### 8. XGBoost

CountVectorizer

grid_xgb = GridSearchCV(XGBClassifier(), param_grid={}, cv=folds, scoring='accuracy')

grid_xgb.fit(train, y_train)

xgb_cv_train_score = grid_xgb.score(train, y_train)
xgb_cv_test_score = accuracy_score(grid_xgb.predict(test), y_test)

xgb_cv_fold_test_mean = grid_xgb.cv_results_['mean_test_score'].mean()
xgb_cv_fold_test_std = grid_xgb.cv_results_['std_test_score'].mean()

print(f'''Best parameters: {grid_xgb.best_params_}
Mean cv test score: {grid_xgb.cv_results_['mean_test_score'].mean()}
Std cv test score: {grid_xgb.cv_results_['std_test_score'].mean()}
Train score: {xgb_cv_train_score}
Test test score: {xgb_cv_test_score}''')

Tfidfvectorizer

grid_xgb.fit(train_tfidf, y_train)

xgb_td_train_score = grid_xgb.score(train_tfidf, y_train)
xgb_td_test_score = accuracy_score(grid_xgb.predict(test_tfidf), y_test)

xgb_td_fold_test_mean = grid_xgb.cv_results_['mean_test_score'].mean()
xgb_td_fold_test_std = grid_xgb.cv_results_['std_test_score'].mean()

print(f'''Best parameters: {grid_xgb.best_params_}
Mean cv test score: {grid_xgb.cv_results_['mean_test_score'].mean()}
Std cv test score: {grid_xgb.cv_results_['std_test_score'].mean()}
Train score: {xgb_td_train_score}
Test test score: {xgb_td_test_score}''')

## Results

import seaborn as sns
import matplotlib.pyplot as plt

from sklearn.metrics import classification_report, confusion_matrix

models = [
    'MultinomialNB_cv',
    'LogisticRegression_cv', 
    'SGD_cv',
    'KNeighbors_cv', 
    'DecisionTree_cv',
    'AdaBoost_cv',
    'GradientBoosting_cv',
    'XGB_cv',
    'MultinomialNB_td',
    'LogisticRegression_td',
    'SGD_td',
    'KNeighbors_td',
    'DecisionTree_td',
    'AdaBoost_td',
    'GradientBoosting_td',
    'XGB_td',
]

train_score_cv = [mnb_train_score_cv, log_cv_train_score, sgd_cv_train_score, knn_cv_train_score, dtc_cv_train_score, ada_cv_train_score, grb_cv_train_score, xgb_cv_train_score]
train_score_td = [mnb_train_score_td, log_td_train_score, sgd_td_train_score, knn_td_train_score, dtc_td_train_score, ada_td_train_score, grb_td_train_score, xgb_td_train_score]

test_score_cv = [mnb_test_score_cv, log_cv_test_score, sgd_cv_test_score, knn_cv_test_score, dtc_cv_test_score, ada_cv_test_score, grb_cv_test_score, xgb_cv_test_score]
test_score_td = [mnb_test_score_td, log_td_test_score, sgd_td_test_score, knn_td_test_score, dtc_td_test_score, ada_td_test_score, grb_td_test_score, xgb_td_test_score]

mean_fold_cv = [mnb_cv_fold_test_mean,log_cv_fold_test_mean,sgd_cv_fold_test_mean, knn_cv_fold_test_mean, dtc_cv_fold_test_mean, ada_cv_fold_test_mean, grb_cv_fold_test_mean, xgb_cv_fold_test_mean]
std_fold_cv = [mnb_cv_fold_test_std,log_cv_fold_test_std, sgd_cv_fold_test_std,knn_cv_fold_test_std,dtc_cv_fold_test_std, ada_cv_fold_test_std, grb_cv_fold_test_std, xgb_cv_fold_test_std] 

mean_fold_td = [mnb_td_fold_test_mean, log_td_fold_test_mean, sgd_td_fold_test_mean,  knn_td_fold_test_mean,  dtc_td_fold_test_mean,  ada_td_fold_test_mean,  grb_td_fold_test_mean, xgb_td_fold_test_mean]
std_fold_td = [mnb_td_fold_test_std,  log_td_fold_test_std, sgd_td_fold_test_std, knn_td_fold_test_std, dtc_td_fold_test_std, ada_td_fold_test_std, grb_td_fold_test_std, xgb_td_fold_test_std] 

pd.DataFrame({'model':models, 'mean fold test score': mean_fold_cv+mean_fold_td,'std fold test score':std_fold_cv+std_fold_td,'train score': train_score_cv + train_score_td, 'test score': test_score_cv + test_score_td})
