==> 7.a.vs_model_selection.ipynb <==
## Model Selection
### Topic (**Inflation**) according to news portal

import warnings
warnings.filterwarnings('ignore')

### Load

import pickle

import pandas as pd
import numpy as np

izq = pickle.load(open('../data/df_clean_izq.pkl','rb'))
der = pickle.load(open('../data/df_clean_der.pkl','rb'))

### Preparte Dataframe

izq = izq[izq.topics == 'inflation']
der = der[der.topics == 'inflation']

izq['bias'] = 'left'
der['bias'] = 'right'

df = izq.loc[:,['bias', 'body']].append(der.loc[:,['bias', 'body']])

### Split Train - Test

from sklearn.model_selection import train_test_split

X = df.body

y = df.bias

X_train, X_test, y_train, y_test = train_test_split(X, y, train_size=0.75, stratify=y)

print(f"Train shape: {X_train.shape[0] / X.shape[0]}\nTest shape: {X_test.shape[0] / X.shape[0]}")

print(f"Target balance\nTrain:\n{y_train.value_counts(normalize=True)}\nTest:\n{y_test.value_counts(normalize=True)}")

### Looking for the model

#### Preparing the input

import re
import unidecode
from nltk.corpus import stopwords
from nltk.tokenize import word_tokenize

stopwords_sp = stopwords.words('spanish')

from sklearn.feature_extraction.text import CountVectorizer, TfidfTransformer

def removal(text):
    text = re.sub(r'(\d|\$|\%|\+)', '', text.lower())
    return re.sub(r'\d+', '', text)

def tokenizer(doc_list):
    terminos = []    
    for texto in doc_list:
        alphanumeric = re.sub(r'([^\s\w]|_)+', '', texto).lower()
        no_accents = unidecode.unidecode(alphanumeric)
        tockens = word_tokenize(no_accents)
        tockens_clean = [tocken for tocken in tockens if tocken not in stopwords_sp and tocken.isalpha()]
    terminos.append(tockens_clean)
    return terminos

#### Folds config

from sklearn.model_selection import StratifiedKFold


folds = StratifiedKFold(n_splits=5, random_state=19, shuffle=True)

#### Calling the models

from sklearn.naive_bayes import MultinomialNB

from sklearn.linear_model import LogisticRegression, SGDClassifier

from sklearn.neighbors import KNeighborsClassifier

from sklearn.tree import DecisionTreeClassifier, plot_tree

from sklearn.ensemble import AdaBoostClassifier, GradientBoostingClassifier

#### Pipeline

from sklearn.pipeline import Pipeline
from sklearn.base import BaseEstimator

class ClfSwitcher(BaseEstimator):

    def __init__(self, estimator = None):
        self.estimator = estimator

    def fit(self, X, y=None, **kwargs):
        self.estimator.fit(X, y)
        return self

    def predict(self, X, y=None):
        return self.estimator.predict(X)

    def predict_proba(self, X):
        return self.estimator.predict_proba(X)

    def score(self, X, y):
        return self.estimator.score(X, y)

pipeline = Pipeline([
    ('cv', CountVectorizer(stop_words=stopwords_sp)),
    ('tfidf', TfidfTransformer()),
    ('clf', ClfSwitcher()),
])

parameters = [
    {
        # vectorizer params
        'cv__max_df': (0.5, 0.75, 1.0),
        'cv__lowercase':  [True],
        'cv__strip_accents':['unicode'], 
        'cv__preprocessor':[removal],
        'cv__ngram_range': [(1,1),(1,2),(2,2)],
        'cv__min_df':[1,2,3],
        # model params
        'clf__estimator': [SGDClassifier()], # SVM if hinge loss / logreg if log loss
        'clf__estimator__penalty': ('l2', 'elasticnet', 'l1'),
        'clf__estimator__max_iter': [50, 80],
        'clf__estimator__tol': [1e-4],
        'clf__estimator__loss': ['hinge', 'log', 'modified_huber'],
    },
    {
        # vectorizer params
        'cv__max_df': (0.5, 0.75, 1.0),
        'cv__lowercase':  [True],
        'cv__strip_accents':['unicode'], 
        'cv__preprocessor':[removal],
        'cv__ngram_range': [(1,1),(1,2),(2,2)],
        'cv__min_df':[1,2,3],
        # model params
        'clf__estimator': [MultinomialNB()],
        'clf__estimator__alpha': (1e-2, 1e-3, 1e-1),
    },
    {
        # vectorizer params
        'cv__max_df': (0.5, 0.75, 1.0),
        'cv__lowercase':  [True],
        'cv__strip_accents':['unicode'], 
        'cv__preprocessor':[removal],
        'cv__ngram_range': [(1,1),(1,2),(2,2)],
        'cv__min_df':[1,2,3],
        # model params
        'clf__estimator': [KNeighborsClassifier()],
        'clf__estimator__n_neighbors': range(1,4),
        'clf__estimator__weights' : ['uniform', 'distance'],
        'clf__estimator__p' : [1, 2, 3]
    },
    {
        # vectorizer params
        'cv__max_df': (0.5, 0.75, 1.0),
        'cv__lowercase':  [True],
        'cv__strip_accents':['unicode'], 
        'cv__preprocessor':[removal],
        'cv__ngram_range': [(1,1),(1,2),(2,2)],
        'cv__min_df':[1,2,3],
        # model params
        'clf__estimator': [LogisticRegression()],
        'clf__estimator__C': [1, 10, 100, 1000],
        'clf__estimator__penalty': ['l1', 'l2',],
        'clf__estimator__solver': ['saga']
    },
    {
        # vectorizer params
        'cv__max_df': (0.5, 0.75, 1.0),
        'cv__lowercase':  [True],
        'cv__strip_accents':['unicode'], 
        'cv__preprocessor':[removal],
        'cv__ngram_range': [(1,1),(1,2),(2,2)],
        'cv__min_df':[1,2,3],
        # model params
        'clf__estimator': [DecisionTreeClassifier()],
        "clf__estimator__criterion" : ["gini", "entropy"],
        "clf__estimator__min_samples_leaf": [5,10,15,20,2], 
        "clf__estimator__max_depth" : [1,2,3,4,5,6,8,9,10,11,12,13,14,15,16,17],
        "clf__estimator__min_samples_split": [2, 3, 4,None]
    },    
]

#### Tuning model

from sklearn.model_selection import GridSearchCV

grid = GridSearchCV(pipeline, parameters, cv=folds, n_jobs=12, return_train_score=False, verbose=3)
grid.fit(X_train, y_train)

#### Report

import seaborn as sns
import matplotlib.pyplot as plt

from sklearn.model_selection import cross_val_score
from sklearn.metrics import classification_report, confusion_matrix, accuracy_score

grid.best_estimator_, grid.best_score_, grid.best_params_

grid_pred = grid.predict(X_test)

accuracy_score(y_test, grid_pred)

print(classification_report(y_test, grid_pred))

confusion_matrix(y_test, grid_pred)
