==> 8.a.sentiment_topic_1.ipynb <==
## Sentiment Analisys
### Input: article lead

## Load Data

import pandas as pd
import numpy as np
import pickle

data = pickle.load(open('data/df_clean_izq.pkl','rb'))

ex_izq = data.loc[data.topics == 'exchange']
ex_izq.shape

data = pickle.load(open('data/df_clean_der.pkl','rb'))

ex_der = data.loc[data.topics == 'exchange']
ex_der.shape

## Text Blob
Translation with TextBlob and googletrans

from textblob import TextBlob
from googletrans import Translator

#### Parts of Speech(POS) Tagging
#### Spanish VS English (Example)
There are evident issues in the tagging process before translation that lead to the imposibility of getting the sentiment scores.

print(f'Input en español:\n{ex_izq.iloc[0].lead}\n')
text = TextBlob(ex_izq.iloc[0].lead)

print(f'POS: {text.tags}\n')

print(f'Sustantivos: {text.noun_phrases}\n')

print(text.sentiment,'\n')

translator = Translator()
translation = translator.translate(ex_izq.iloc[0].lead, src='es', dest='en')

print(f'Traducción:\n{translation.text}\n')
text = TextBlob(translation.text)

print(f'POS: {text.tags}\n')

print(f'Sustantivos: {text.noun_phrases}\n') 

print(text.sentiment)

print(f'Input en español: {ex_der.iloc[0].lead}\n')
text = TextBlob(ex_der.iloc[0].lead)

print(f'POS: {text.tags}\n')

print(f'Sustantivos: {text.noun_phrases}\n')

print(text.sentiment,'\n')

translator = Translator()
translation = translator.translate(ex_der.iloc[0].lead, src='es', dest='en')

print(f'Traducción: {translation.text}\n')
text = TextBlob(translation.text)

print(f'POS: {text.tags}\n')

print(f'Sustantivos: {text.noun_phrases}\n') 

print(text.sentiment)

### Datasets - Sentiment Analisys

#### 1. "derecha diario"

Text blob with google translator

translator = Translator()
pol_t = lambda x: np.around(TextBlob(translator.translate(x, src='es', dest='en').text).sentiment.polarity, 3)
sub_t = lambda x: np.around(TextBlob(translator.translate(x, src='es', dest='en').text).sentiment.subjectivity, 3)

ex_der['polarity'] = ex_der.lead.apply(pol_t)
ex_der['subjectivity'] = ex_der.lead.apply(sub_t)

print(f'Polarity: {ex_der.polarity.mean()}\nSubjectivity: {ex_der.subjectivity.mean()}')

Text blob wit textblob translator

pol = lambda x: np.around(TextBlob(x).translate(to='en').sentiment.polarity, 3)
sub = lambda x: np.around(TextBlob(x).translate(to='en').sentiment.subjectivity, 3)

ex_der['polarity'] = ex_der.lead.apply(pol)
ex_der['subjectivity'] = ex_der.lead.apply(sub)

print(f'Polarity: {ex_der.polarity.mean()}\nSubjectivity: {ex_der.subjectivity.mean()}')

#### 2. "izquierda diario"

ex_izq['polarity'] = ex_izq.lead.apply(pol)
ex_izq['subjectivity'] = ex_izq.lead.apply(sub)

print(f'Polarity: {ex_izq.polarity.mean()}\nSubjectivity: {ex_izq.subjectivity.mean()}')

ex_izq['polarity'] = ex_izq.lead.apply(pol_t)
ex_izq['subjectivity'] = ex_izq.lead.apply(sub_t)

print(f'Polarity: {ex_izq.polarity.mean()}\nSubjectivity: {ex_izq.subjectivity.mean()}')
