==> 8.b.sentiment_topic_2.ipynb <==
## Sentiment Analisys

### Load Data

import pandas as pd
import numpy as np
import pickle

data = pickle.load(open('data/df_clean_izq.pkl','rb'))

inf_izq = data.loc[data.topics == 'inflation']
inf_izq.shape

data = pickle.load(open('data/df_clean_der.pkl','rb'))

inf_der = data.loc[data.topics == 'inflation']
inf_der.shape

### Parts of Speech (POS) Tagging and Sentiment Analysis

- CC coordinating conjunction
- CD cardinal digit
- DT determiner
- EX existential there
- FW foreign word
- IN preposition/subordinating conjunction
- JJ This NLTK POS Tag is an adjective (large)
- JJR adjective, comparative (larger)
- JJS adjective, superlative (largest)
- LS list market
- MD modal (could, will)
- NN noun, singular (cat, tree)
- NNS noun plural (desks)
- NNP proper noun, singular (sarah)
- NNPS proper noun, plural (indians or americans)
- PDT predeterminer (all, both, half)
- POS possessive ending (parent\ 's)
- PRP personal pronoun (hers, herself, him,himself)
- PRPS possessive pronoun (her, his, mine, my, our )
- RB adverb (occasionally, swiftly)
- RBR adverb, comparative (greater)
- RBS adverb, superlative (biggest)
- RP particle (about)
- TO infinite marker (to)
- UH interjection (goodbye)
- VB verb (ask)
- VBG verb gerund (judging)
- VBD verb past tense (pleaded)
- VBN verb past participle (reunified)
- VBP verb, present tense not 3rd person singular(wrap)
- VBZ verb, present tense with 3rd person singular (bases)
- WDT wh-determiner (that, what)
- WP wh- pronoun (who)
- WRB wh- adverb (how) 

import nltk

from nltk.sentiment import SentimentIntensityAnalyzer

# nltk.download('vader_lexicon')

from textblob import TextBlob
from googletrans import Translator

### NLTK vs TextBlob (tagging)

Import tokens for nltk tagging

token_df = pd.read_pickle('tokens/topic_tokens_izq.pkl')

token_df = token_df[token_df.topics == 'inflation']

token_df.drop(['topics'], axis=1, inplace=True)

# made a corpus from tokens in df
corpus = []
for row in range(len(token_df)):
    corpus.append(token_df.iloc[row,(token_df.iloc[row] == 1).values].index.tolist())
len(corpus)

nltk (tokens) vs textblob (text)

display( dict(nltk.pos_tag(corpus[0])).get('comision'),
        dict(nltk.pos_tag(corpus[0])).get('nacional'),
        dict(nltk.pos_tag(corpus[0])).get('valores'),
        dict(nltk.pos_tag(corpus[0])).get('cantidad'),
        dict(nltk.pos_tag(corpus[0])).get('tener'),
        dict(nltk.pos_tag(corpus[0])).get('venderlo'),
       )

text = TextBlob(inf_izq.iloc[0, -2])
print(text.tags[:60])

### Text Blob and language
Translation with TextBlob and googletrans

#### Spanish VS English (Example)
There are evident issues in the tagging process before translation that lead to the imposibility of getting the sentiment scores.

# Spanish
print(f'Input en español:\n{inf_izq.iloc[0].lead}\n')

text = TextBlob(inf_izq.iloc[0].lead)
print(f'POS:\n{text.tags}\n')

print(f'Sustantivos:\n{text.noun_phrases}\n')

print(text.sentiment,'\n')

# English
translator = Translator()
translation = translator.translate(inf_izq.iloc[0].lead, src='es', dest='en')

print(f'Traducción:\n{translation.text}\n')

text = TextBlob(translation.text)
print(f'POS:\n{text.tags}\n')

print(f'Sustantivos:\n{text.noun_phrases}\n') 

print(text.sentiment)

print(f'Input en español: {inf_der.iloc[0].lead}\n')
text = TextBlob(inf_der.iloc[0].lead)

print(f'POS:\n{text.tags}\n')

print(f'Sustantivos:\n{text.noun_phrases}\n')

print(text.sentiment,'\n')

translator = Translator()
translation = translator.translate(inf_der.iloc[0].lead, src='es', dest='en')

print(f'Traducción: {translation.text}\n')
text = TextBlob(translation.text)

print(f'POS:\n{text.tags}\n')

print(f'Sustantivos:\n{text.noun_phrases}\n') 

print(text.sentiment)

### Datasets - Sentiment Analisys

#### NLTK
with nltk VADER (Valence Aware Dictionary and sEntiment Reasoner)

# nltk sentiment
sia = SentimentIntensityAnalyzer()
print(f'Example: {sia.polarity_scores(inf_izq.iloc[0, :].body)}')

def polarity(df):
    nltk_pol = lambda x: sia.polarity_scores(x)
    df_pol = df.body.apply(nltk_pol)
    neg = 0; pos = 0; neu = 0
    for row in list(df_pol.values):
        pos += row['pos']
        neg += row['neg']
        neu += row['neu']
        return {'neg': neg,
                'pos': pos,
                'neu': neu
               }

#### 1. "derecha diario"

nltk

print(f'NLTK\n{polarity(inf_der)}')

Text blob with google translator

translator = Translator()
pol_t = lambda x: np.around(TextBlob(translator.translate(x, src='es', dest='en').text).sentiment.polarity, 3)
sub_t = lambda x: np.around(TextBlob(translator.translate(x, src='es', dest='en').text).sentiment.subjectivity, 3)

inf_der['polarity'] = inf_der.lead.apply(pol_t)
inf_der['subjectivity'] = inf_der.lead.apply(sub_t)

print(f'TextBlob\nPolarity: {inf_der.polarity.mean()}\nSubjectivity: {inf_der.subjectivity.mean()}')

Text blob with textblob translator

# pol = lambda x: np.around(TextBlob(x).translate(to='en').sentiment.polarity, 3)
# sub = lambda x: np.around(TextBlob(x).translate(to='en').sentiment.subjectivity, 3)

# inf_der['polarity'] = inf_der.lead.apply(pol)
# inf_der['subjectivity'] = inf_der.lead.apply(sub)

# print(f'Polarity: {inf_der.polarity.mean()}\nSubjectivity: {inf_der.subjectivity.mean()}')

#### 2. "izquierda diario"

nltk

print(f'NLTK\n{polarity(inf_izq)}')

Text blob with google translator

#inf_izq['polarity'] = inf_izq.lead.apply(pol)
#inf_izq['subjectivity'] = inf_izq.lead.apply(sub)

# print(f'Polarity: {inf_izq.polarity.mean()}\nSubjectivity: {inf_izq.subjectivity.mean()}')

inf_izq['polarity'] = inf_izq.lead.apply(pol_t)
inf_izq['subjectivity'] = inf_izq.lead.apply(sub_t)

print(f'TextBlob\nPolarity: {inf_izq.polarity.mean()}\nSubjectivity: {inf_izq.subjectivity.mean()}')
