==> 9.gensim_topic_modeling.ipynb <==
import pandas as pd
import numpy as np
import pickle

### Data

Izquierda Diario

df_izq = pickle.load(open('data/df_clean_izq.pkl','rb'))

df_inflation_izq = df_izq[df_izq.topics == 'inflation']

df_exchange_izq = df_izq[df_izq.topics == 'exchange']

corpus_izq = df_inflation_izq.append(df_exchange_izq)

Derecha Diario

df_der = pickle.load(open('data/df_clean_der.pkl','rb'))

df_inflation_der = df_der[df_der.topics == 'inflation']

df_exchange_der = df_der[df_der.topics == 'exchange']

corpus_der = df_inflation_der.append(df_exchange_der)

### Tokenization

import re
import unidecode

from nltk.corpus import stopwords
import spacy
from spacy.lang.es import Spanish

parser = Spanish()
stopwords_sp = stopwords.words('spanish') + ['si','dia', 'vez', 'fin','dias', 'dos', 'mas','ano', 'asi','puede', 'mes','incluso']

def tokenizer(texto):

    alphanumeric = re.sub(r'([^\s\w]|_)+', '', texto).lower()
    no_accents = unidecode.unidecode(alphanumeric)
    
    tokens = parser(no_accents)
    
    tokens_list = [str(token) for token in tokens]
    
    tokens_clean = [token for token in tokens_list if token not in stopwords_sp and token.isalpha()]

    return tokens_clean

tokens_izq = [tokenizer(text) for text in corpus_izq.body]

tokens_der = [tokenizer(text) for text in corpus_der.body]

### Gensim

Give ids to tokens with gensim Dictionary

# import logging
# logging.basicConfig(format='%(asctime)s : %(levelname)s : %(message)s', level=logging.INFO)

from gensim.corpora import Dictionary

# tokens input
dictionary_izq = Dictionary(tokens_izq)
dictionary_izq.filter_extremes(no_below=2, no_above=0.97, keep_n=None)

# convert each document into the gensim bag-of-words
corpus_izq = [dictionary_izq.doc2bow(text) for text in tokens_izq]

# tokens input
dictionary_der = Dictionary(tokens_der)
dictionary_der.filter_extremes(no_below=2, no_above=0.97, keep_n=None)

# convert each document into the gensim bag-of-words
corpus_der = [dictionary_der.doc2bow(text) for text in tokens_der]

### Latent Dirichlet Allocation (LDA)
Check the unsupervised word classification using LDA

from gensim import matutils, models
import scipy.sparse

Izquierda Diario

# make a dictionary with the ids of each word assigned by Gensim dictionary method

id_word_izq = dict((v, k) for k, v in dictionary_izq.token2id.items())

# we already know that there are two topics

lda_izq = models.LdaModel(corpus=corpus_izq, id2word=id_word_izq, num_topics=2, passes=20, random_state= 30)

lda_izq.print_topics()

import pyLDAvis.gensim

lda_display = pyLDAvis.gensim.prepare(lda_izq, corpus_izq, dictionary_izq, sort_topics=False)

pyLDAvis.display(lda_display)

Derecha Diario

# make a dictionary with the ids of each word assigned by Gensim dictionary method

id_word_der = dict((v, k) for k, v in dictionary_der.token2id.items())

# we already know that there are two topics

lda_der = models.LdaModel(corpus=corpus_der, id2word=id_word_der, num_topics=2, passes=20, random_state= 30)

lda_der.print_topics()

lda_display = pyLDAvis.gensim.prepare(lda_der, corpus_der, dictionary_der, sort_topics=False)
pyLDAvis.display(lda_display)

### Filtering by tags

import nltk

# tag is not accurate
nltk.pos_tag(tokens_izq[0][:10])

from polyglot.text import Text
!polyglot download embeddings2.es pos2.es

def noun_only(tokens):
    nouns = []
    for token in tokens:
        word_t = Text(token, hint_language_code='es')
        if word_t.pos_tags[0][1] == 'NOUN':
            nouns.append(word_t.pos_tags[0][0])
    return nouns

tokens_nouns_izq = [noun_only(token) for token in tokens_izq]
tokens_nouns_der = [noun_only(token) for token in tokens_der]

### New Noun Dictionary + LDA

# tokens input
dictionary_izq = Dictionary(tokens_nouns_izq)
dictionary_izq.filter_extremes(no_below=2, no_above=0.97, keep_n=None)

# convert each document into the gensim bag-of-words
corpus_izq = [dictionary_izq.doc2bow(text) for text in tokens_nouns_izq]

# tokens input
dictionary_der = Dictionary(tokens_nouns_der)
dictionary_der.filter_extremes(no_below=2, no_above=0.97, keep_n=None)

# convert each document into the gensim bag-of-words
corpus_der = [dictionary_der.doc2bow(text) for text in tokens_nouns_der]

Izquierda Diario

id_word_izq = dict((v, k) for k, v in dictionary_izq.token2id.items())

lda_izq = models.LdaModel(corpus=corpus_izq, id2word=id_word_izq, num_topics=2, passes=20, random_state= 30)

lda_izq.print_topics()

lda_display = pyLDAvis.gensim.prepare(lda_izq, corpus_izq, dictionary_izq, sort_topics=False)
pyLDAvis.display(lda_display)

Derecha Diario

id_word_der = dict((v, k) for k, v in dictionary_der.token2id.items())

lda_der = models.LdaModel(corpus=corpus_der, id2word=id_word_der, num_topics=2, passes=20, random_state= 30)

lda_der.print_topics()

lda_display = pyLDAvis.gensim.prepare(lda_der, corpus_der, dictionary_der, sort_topics=False)
pyLDAvis.display(lda_display)
